Refactor tests and TestConstants
